### 2020-08-14 #2572 @whedon
[PRE REVIEW]: SLAM Toolbox: SLAM for the dynamic world
**Submitting author:** @stevemacenski (<a href="http://orcid.org/0000-0003-1090-7733">Steven Macenski</a>)
**Repository:** <a href="https://github.com/SteveMacenski/slam_toolbox" target ="_blank">https://github.com/SteveMacenski/slam_toolbox</a>
**Version:** 2.3.0
