### 2020-08-14 #2572 @SteveMacenski
> Can't find any papers to compile :-(

I'm not sure why it couldn't find the file, it built fine with the preview tool. Its on the `joss` branch.

carlosjoseRG, mosteo, or traversaro could make good editors (all listing C++ and robotics)
