### 2020-08-18 2cdee7a
Update bug_report.md
