### 2020-08-19 762b904
fix for typo in distance calculation (getX-getY) (#225)
* fix for typo in distance calculation (getX-getY)

* no todo

Co-authored-by: Nikolas Engelhard <nikolas.engelhard@motius.de>
