### 2020-08-19 ac1d5a2
fixing dist calc on noetic (#283)
