### 2020-08-19 e94df36
fixing dist calc in foxy (#282)
