### 2020-09-09 8b8b7ef
Handle case for lidar with 1 degree resolution that goes from [-179,180] with 360 readings (#289)
