### 2020-09-09 d9a8a86
Handle case for lidar with 1 degree resolution that goes from [-179,180] with 360 readings (which RPLIDAR does in angle_compensate mode).  Works with patched RPLIDAR driver here: https://github.com/allenh1/rplidar_ros/pull/20 (#288)
