### 2020-09-11 6c2afcd
[Noetic] Fix build failure (#291)
Adds working dependency on slam_toolbox_msgs to
slam_toolbox_rviz to ensure correct build order.
