### 2020-09-16 #2572 @bstabler
Hi @SteveMacenski - I finally read the paper.  I have a couple small comments on the paper while we look for reviewers:
  - there is a typo "proided"
  - an explicit Statement of Need section needs to be added
  - the Configuration section of the paper is not needed and should probably instead be in the online and/or package documentation

@carlosjoseRG, @mosteo, @muhaochen, @breznak, @jcarpent, or @traversaro - are you interested in reviewing this paper?
