### 2020-09-16 #2572 @SteveMacenski
Hi,

Thanks for the review!
- Typo fixed
- I'd be happy to. Should this be another section in the paper? I based my format on [this](https://joss.theoj.org/papers/10.21105/joss.01531) and [this](https://joss.theoj.org/papers/10.21105/joss.00456) paper in JOSS in the robotics space, which have that built into the summary. I can reiterate the statement of need in a separate section if that would be preferable. Let me know what you think.
- That content is also available in the readme. I figured it would be good to add here as well, even if not required, to add some more context to the reader. I can remove it if you like, but I think it provides some useful context if this is their first introduction to the work in reading the paper.
