### 2020-09-17 #2572 @bstabler
Yes, it should now be an explicit section of the paper, see https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain.  For the second item, I think it's too much detail for this high level introduction / summary.  Thanks.
