### 2020-09-17 ea6ee14
remove configuration section
