### 2020-09-17 #2572 @SteveMacenski
Ok,

- Statement of need section added https://github.com/SteveMacenski/slam_toolbox/commit/efb33ec1afe4bfee7e26624a0c3f48ccfb8ec721
- Configuration section removed https://github.com/SteveMacenski/slam_toolbox/commit/ea6ee14bce12885024342ce598c8b15a375d50bb
