### 2020-09-23 7bc6dd3
adding fix for 360 inclusive bound lidars
