### 2020-09-23 08f7674
adding fix for 360 inclusive bound lidars
