### 2020-09-23 add9969
adding fix for 360 inclusive bound lidars
