### 2020-10-01 1ab079e
Update laser_utils.cpp
