### 2020-10-01 6a323fb
Update laser_utils.cpp
