### 2020-10-01 c2479cf
Update laser_utils.cpp
