### 2020-10-01 77337f4
[Foxy] adding fix for 360 inclusive bound lidars (#296)
* adding fix for 360 inclusive bound lidars

* adding std::round

* Update laser_utils.cpp

* Update laser_utils.cpp
