### 2020-10-01 44c9f84
[melodic] adding fix for 360 inclusive bound lidars (#298)
* adding fix for 360 inclusive bound lidars

* Update laser_utils.cpp

* Update laser_utils.cpp
