### 2020-10-08 75d0c86
update CMakeLists for correct boost / flags
