### 2020-10-08 5364ba1
update CMakeLists for correct boost / flags (#301)
