### 2020-10-14 0cf69f2
copying timestamp from ros-msg to LocalizedRangeScan (#302)
Co-authored-by: Nikolas Engelhard <nikolas.engelhard@motius.de>
