### 2020-10-25 #2572 @arfon
@onlytailei @pifon2a @ojura - would you be interested in reviewing this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under review is [_SLAM Toolbox: SLAM for the dynamic world_](https://github.com/SteveMacenski/slam_toolbox), the paper we'd ask you to review is available [here](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02572/joss.02572/10.21105.joss.02572.pdf).
