### 2020-10-27 #2572 @carlosjoserg
> @SteveMacenski - apologies for the very slow progress here. I'll be taking over to help out Ben here.
>
>  @mosteo & @carlosjoserg - would you be interested in reviewing this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> The software under review is [_SLAM Toolbox: SLAM for the dynamic world_](https://github.com/SteveMacenski/slam_toolbox), the paper we'd ask you to review is available [here](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02572/joss.02572/10.21105.joss.02572.pdf).

Sorry, I'm just seeing this thread now. I'm not a SLAM expert, I come more from the manipulation world. If you are interested in me reviewing the soft/paper, let me know.
