+++ launch/merge_maps_kinematic_launch.py
@@ -6,7 +6,7 @@ def generate_launch_description():
     return LaunchDescription([
         launch_ros.actions.Node(
             package='slam_toolbox',
-            node_executable='merge_maps_kinematic',
+            executable='merge_maps_kinematic',
             name='slam_toolbox',
             output='screen'
         )
