+++ launch/online_sync_launch.py
@@ -19,7 +19,7 @@ def generate_launch_description():
           {'use_sim_time': use_sim_time}
         ],
         package='slam_toolbox',
-        node_executable='sync_slam_toolbox_node',
+        executable='sync_slam_toolbox_node',
         name='slam_toolbox',
         output='screen')
 
