### 2020-12-18 #2783 @mosteo
Opened a PR with proofreading fixes: SteveMacenski/slam_toolbox#317
