### 2020-12-18 #2783 @mosteo
Installation verified by installing the `ros-foxy-slam-toolbox` package.
