### 2020-12-18 SteveMacenski/slam_toolbox#318 @mosteo
[JOSS review] Example usage
Hello, Steve.

I'm one of the reviewers at https://github.com/openjournals/joss-reviews/issues/2783.

I'm opening this issue in regard to the point:
- [ ] **Example usage**: Do the authors include examples of how to use the software?

Although there are a few videos and captures, the only hands-on info I see in the README is the line:

> You can run via ros2 launch slam_toolbox online_sync_launch.py

which, by itself, doesn't bring up a complete example.

The page https://github.com/SteveMacenski/slam_toolbox/wiki does not contain one either.

Is https://navigation.ros.org/tutorials/docs/navigation2_with_slam.html the intended complete example?
