### 2020-12-18 SteveMacenski/slam_toolbox#318 @SteveMacenski
Hi, yes we have example usage tutorials
- https://navigation.ros.org/tutorials/docs/navigation2_with_slam.html (how to do slam while navigating)
- http://wiki.ros.org/Robots/ARI/Tutorials/Navigation/Mapping (an unrelated company created a tutorial for their customers based on this work)
- A live set of demos in the ROSCon talk regarding it https://vimeo.com/378682207 that goes into alot more details on capabilities and configuration (also linked to in readme)

The Nav2 website tutorial is a canonical "getting started" task that all new developers in ROS go through. Since it leverages a bunch of ROS standards, its a drop in replacement for others in the ecosystem.

While this is a stand alone piece of software, you have to have inputs coming in from sensors (real or simulated) to have it do anything of interest. The required inputs are described https://github.com/SteveMacenski/slam_toolbox#subscribed-topics and these are well-established standards for the ROS open-source community that all SLAM and localization implementations comply to.

Launching any ROS node in isolation probably won't do anything, you need the computation graph around it (like the gazebo simulation and navigation launch in the tutorial) to give the inputs to it.
