### 2020-12-21 SteveMacenski/slam_toolbox#320 @carlosjoserg
JOSS Review
Hi @SteveMacenski

I'm reviewing your submission to JOSS _SLAM Toolbox: SLAM for the dynamic world_

This is a first round of issues on the checklists proposed in https://github.com/openjournals/joss-reviews/issues/2783:

1. > **Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?

By quick inspection, I see only [this file](https://github.com/SteveMacenski/slam_toolbox/blob/foxy-devel/test/lifelong_metrics_test.cpp) containing automated tests.

But also see that it is comented in the [CMakeLists.txt](https://github.com/SteveMacenski/slam_toolbox/blob/foxy-devel/CMakeLists.txt#L169-L177), at least for the `foxy-devel` and the [`melodic-devel`](https://github.com/SteveMacenski/slam_toolbox/blob/melodic-devel/slam_toolbox/CMakeLists.txt#L154-L160) branches.

Is there any reason for this non-homogenous testing across the branches?

2. > **Example usage**: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I saw the thread https://github.com/SteveMacenski/slam_toolbox/issues/318

Having to set and run a gazebo world is too much of a burden to being able to run an example. Consider having just the required topics (for instance, as a rosbag) to emulate and run an example.

3. > **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

(1) I could not find any contribution guidelines (see for example [MoveIt!](https://moveit.ros.org/documentation/contributing/) or [MongoCPPDriver](https://github.com/mongodb/mongo-cxx-driver/blob/master/CONTRIBUTING.md)).

Looking at the [branches](https://github.com/SteveMacenski/slam_toolbox/branches) does not help either. I'd find helpful to state the branching strategy (see again [MoveIt](https://github.com/ros-planning/moveit#branches-policy) or [MongoCPPDriver](https://github.com/mongodb/mongo-cxx-driver#branches---releasesstable-versus-master))

(2) Github templates.. OK

(3) I could not find an explicit way to do that. (e.g. Github Issue, ROS Answers, Discourse, Slack, Email, etc.)
