### 2020-12-21 SteveMacenski/slam_toolbox#320 @SteveMacenski
Hi, thanks for your time and notes.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

We have automated testing suites as part of the build farm as well (linting, code styling, static analysis, etc) that are done behind the scenes and pass for every PR, merge, and distribution release. These are _way_ more complete in ROS2. We meet that requirement as a requirement of the first and second: we have strong static analysis and linting automated test suites and manual steps as part of the tutorials you linked to as manual steps for verification on live data streams.

I definitely know we need more automated test coverage, but I think we squeak by on the actual requirement and the spirit of it. More tests are planned.

> Having to set and run a gazebo world is too much of a burden to being able to run an example. Consider having just the required topics (for instance, as a rosbag) to emulate and run an example

Not a bad idea, however I'm not able to share any of my datasets publicly. Keep in mind we do having the required topics, see the https://github.com/SteveMacenski/slam_toolbox#subscribed-topics section of the readme. So I think we do have what you mean, just lack a specific `rosbag play...` CLI command embedded inline which can obvious vary alot by what bag you're trying to run.

Perhaps an aside, but I disagree that its "too much of a burden", that's subjective to your target audience. As this is a production-grade SLAM solution, its the expectation of a minimum viable robot system in order to even generate the data that any SLAM system requires. If those can't be provided, we do offer our tutorials which documents the steps to do so in gazebo so that anyone and everyone can have an example use. I'd also argue that installing ROS is quite a burden to test this, but that's totally unavoidable. Niche things like like naturally have a higher barrier to entry than a basic importable python library.

I think we meet the requirement of having example usages (even more than 1), though a rosbag of REP-103 compliant data would work fine, I have plenty that I run it over to verify major PR changes. The average new user doesn't come loaded up with datasets so being able to run a real simulation with real-live simulated data using the canonical demo of ROS is a far more practical and meeting the "Ideally..." note in the requirement list.

> Contribute to the software

You are right, I overlooked that, thank you. I edited the default branch readme file to give that context. File tickets or email me; then filling in templates and we'll get back to users. https://github.com/SteveMacenski/slam_toolbox#support-and-contribution

> Report issues or problems with the software

I think we both agree this was met?

> Seek support

You are right, I overlooked that, thank you. I edited the default branch readme file to give that context. Config / help -> ROS Answers; issues/features -> issues. Second part also makes more explicit point 2.  https://github.com/SteveMacenski/slam_toolbox#support-and-contribution
