### 2020-12-22 2b73ac0
fix mutually exclusive mapping start pose (#319)
