### 2020-12-28 ea83408
Fix deprecated rclcpp::Duration() calls
