### 2020-12-30 SteveMacenski/slam_toolbox#318 @mosteo
To fulfill the JOSS guideline, these tutorials should be linked from the README.

Also, as I saw in https://github.com/SteveMacenski/slam_toolbox/issues/320, and in particular about

> Perhaps an aside, but I disagree that its "too much of a burden"

I don't consider installing ROS2+Gazebo excessive, but I think that at least a pointer to a complete, self-contained example should be provided in the README. One that when followed to the letter from start to finish, having only cursory familiarity with a similar framework (e.g., Orocos) would get you a working demo in simulation. I see most of the pieces are here and there, but for some glue.

https://navigation.ros.org/tutorials/docs/navigation2_with_slam.html is almost there, provided that installation instructions were complete (some necessary packages are not mentioned) and it could be entirely done in simulation.
