### 2021-01-04 70812be
do not start with pose graph on bad configurations
