### 2021-01-04 3c82f35
use PARAMETER_NOT_SET to test mutually exclusive parameters
