+++ src/slam_toolbox_common.cpp
@@ -300,12 +300,13 @@ bool SlamToolbox::shouldStartWithPoseGraph(
 {
   // if given a map to load at run time, do it.
   this->declare_parameter("map_file_name", std::string(""));
-  this->declare_parameter("map_start_pose");
-  this->declare_parameter("map_start_at_dock");
+  auto read_pose_param = this->declare_parameter("map_start_pose");
+  auto start_at_dock_param = this->declare_parameter("map_start_at_dock");
   filename = this->get_parameter("map_file_name").as_string();
   if (!filename.empty()) {
     std::vector<double> read_pose;
-    if (this->get_parameter("map_start_pose", read_pose)) {
+    if (read_pose_param.get_type() != rclcpp::ParameterType::PARAMETER_NOT_SET) {
+      read_pose = read_pose_param.get<std::vector<double>>();
       start_at_dock = false;
       if (read_pose.size() != 3) {
         RCLCPP_ERROR(get_logger(), "LocalizationSlamToolbox: Incorrect "
@@ -319,8 +320,8 @@ bool SlamToolbox::shouldStartWithPoseGraph(
         pose.y = read_pose[1];
         pose.theta = read_pose[2];
       }
-    } else if (this->get_parameter("map_start_at_dock", start_at_dock)) {
-      start_at_dock = this->get_parameter("map_start_at_dock").as_bool();
+    } else if (start_at_dock_param.get_type() != rclcpp::ParameterType::PARAMETER_NOT_SET) {
+      start_at_dock = start_at_dock_param.get<bool>();
     } else {
       RCLCPP_ERROR(get_logger(), "LocalizationSlamToolbox: Map starting "
           "pose not specified. Set either map_start_pose or map_start_at_dock.");
