### 2021-01-04 SteveMacenski/slam_toolbox#318 @SteveMacenski
> To fulfill the JOSS guideline, these tutorials should be linked from the README.

Fair enough, has been added: https://github.com/SteveMacenski/slam_toolbox/blob/foxy-devel/README.md

> provided that installation instructions were complete (some necessary packages are not mentioned)

I'm not sure what you mean by that. The installation instructions should be complete, many users use that to get started and I haven't heard an issue in quite awhile on that.

I'm not sure I totally grasp your other comment. Is there an action item I need to take? I added it to the readme which I think fulfills your first comment but I don't get the Orocos comment.
