### 2021-01-05 4ddc571
fix mutually exclusive mapping start pose (#319) (#321)
* fix mutually exclusive mapping start pose (#319)

* do not start with pose graph on bad configurations

* use PARAMETER_NOT_SET to test mutually exclusive parameters

* rename parameters
