### 2021-01-16 #2783 @arfon
:wave: @mosteo, @carlosjoserg - happy new year. Just checking in on your reviews here? It looks like @SteveMacenski responded to the questions you asked in https://github.com/SteveMacenski/slam_toolbox/issues/318 and https://github.com/SteveMacenski/slam_toolbox/issues/320 ?
