### 2021-01-25 SteveMacenski/slam_toolbox#318 @mosteo
Sorry for the absence.

Yes, you're correct, I had missed some context with my previous message. I meant that the simple tutorial you link to is not a starting point but another one in a series for the ARI robot, which may be confusing. Also, if my reading is correct, that's for ROS1.

Ideally, a self-contained tutorial for ROS2 would be optimal, with a simpler simulated robot. But I guess that's icing on the cake.
