### 2021-01-25 #2783 @mosteo
I'm back on track, sorry for the delay. @arfon, in regard to the verification of the functionality/performance, this is intended to be verified by the reviewers? Or would evidence of usage by third parties be enough?
