### 2021-01-25 #2783 @SteveMacenski
A few examples from people that aren't me and I've had no contact with (googled and found for reference) if that helps
- https://www.youtube.com/watch?v=ftfMsA-UykQ / https://www.notion.so/Tutorial-SLAM-toolbox-aac021ec21d24f898ce230c19def3b7b
- https://www.youtube.com/watch?v=s16269kol5M
- https://www.youtube.com/watch?v=Cgcl3LcFnEs
- http://www.robotandchisel.com/2020/08/19/slam-in-ros2/ (though note I do know Mike but I didn't ask/inform him on this)
- https://msadowski.github.io/hands-on-with-slam_toolbox/
- https://blog.pal-robotics.com/aris-wiki-ros-tutorials-on-slam/
