### 2021-01-26 #2783 @arfon
> I'm back on track, sorry for the delay. @arfon, in regard to the verification of the functionality/performance, this is intended to be verified by the reviewers? Or would evidence of usage by third parties be enough?

We would definitely prefer both reviewers to verify the functionality claims (performance is sometimes more challenging for folks).
