### 2021-01-26 #2783 @mosteo
Understood. Then, I'm going to throw a ball to @SteveMacenski : I don't currently have access to my labs robots due to covid. Also, I'm exclusively using ROS2 these days. This brings me back to the issue of beginner tutorials. If I go to https://navigation.ros.org/tutorials/docs/navigation2_with_slam.html, I read:

> we will use the turtlebot3. If you have another robot, replace with suitable instructions

I would really like to see there, instead of "replace with suitable", something like "use this previous tutorial to set up a simulated robot" or similar. Related to my earlier comment that people not current with ROS2 may hit a bump there if interested in a "quick" test.
