### 2021-01-26 #2783 @SteveMacenski
Ok, makes sense - do you have a ROS2 bag file you can run it over? I'd be happy to share with you a ROS(1) bag file I have for the purposes of review and a set of instructions to reproduce, but unfortunately I don't have any ROS2 bag files (as you're aware until exceptionally recently, ROS2 bag has been very flaky). ROS 1 would be easier to see everything since that's what this article was written on but lets see what we can work out in ROS2.

---

Another option is as you've found in the tutorial, if you're OK installing Nav2 to run our canonical [getting started demo](https://navigation.ros.org/getting_started/index.html) then one of the parameters I have conveniently provided is [`slam`](https://github.com/ros-planning/navigation2/blob/main/nav2_bringup/bringup/launch/tb3_simulation_launch.py#L35). So

```
ros2 launch nav2_bringup tb3_simulation_launch.py slam:=true
```

will get you what you need. It will launch a TB3 in a sandbox world that you can initialize the pose with the rviz "Pose2D" tool and then request navigation goals with "goal pose" tool. You'll see the map update as you move.The SLAM specific tutorial is meant to be more abstract and separate out the navigation from the simulation from the SLAM since that tutorial is written with "bring your own robot" in mind - in which case using our one-stop-shop launch file `tb3_simulation_launch.py` isn't appropriate.

>  we will use the turtlebot3. If you have another robot, replace with suitable instructions

Is to mean your own robot state publisher, hardware / simulation interface, and any other robot-specific needs. It's hard to fully articulate the broad range of things that a particular company / robot might require, so we keep it abstract. It could be as little as the robot state publisher with URDF and drivers, but frequently its alot more.
