### 2021-02-03 d5dcd6b
fix the issue (#198) that loop closures can rotate the map 180° if the lidar is mounted backwards
Set the edges (constraints) based on the robot base frame (base_footprint) instead of sensor frame (laser).
