### 2021-02-05 5cb6690
add function GetCorrectedAt to simplify getting robot pose
