### 2021-02-05 9b8f6f0
add function GetCorrectedAt to simplify getting robot pose
