### 2021-02-05 5d637e5
add function GetCorrectedAt to simplify getting robot pose
