+++ lib/karto_sdk/include/karto_sdk/Karto.h
@@ -5538,17 +5538,7 @@ public:
    */
   void SetSensorPose(const Pose2 & rScanPose)
   {
-    Pose2 deviceOffsetPose2 = GetLaserRangeFinder()->GetOffsetPose();
-    kt_double offsetLength = deviceOffsetPose2.GetPosition().Length();
-    kt_double offsetHeading = deviceOffsetPose2.GetHeading();
-    kt_double angleoffset = atan2(deviceOffsetPose2.GetY(), deviceOffsetPose2.GetX());
-    kt_double correctedHeading = math::NormalizeAngle(rScanPose.GetHeading());
-    Pose2 worldSensorOffset =
-      Pose2(offsetLength * cos(correctedHeading + angleoffset - offsetHeading),
-        offsetLength * sin(correctedHeading + angleoffset - offsetHeading),
-        offsetHeading);
-
-    m_CorrectedPose = rScanPose - worldSensorOffset;
+    m_CorrectedPose = GetCorrectedAt(rScanPose);
 
     Update();
   }
@@ -5562,6 +5552,25 @@ public:
   {
     return Transform(rPose).TransformPose(GetLaserRangeFinder()->GetOffsetPose());
   }
+
+    /**
+     * @brief Computes the pose of the robot if the sensor were at the given pose
+     * @param sPose sensor pose
+     * @return robot pose
+     */
+    inline Pose2 GetCorrectedAt(const Pose2& sPose) const
+    {
+      Pose2 deviceOffsetPose2 = GetLaserRangeFinder()->GetOffsetPose();
+      kt_double offsetLength = deviceOffsetPose2.GetPosition().Length();
+      kt_double offsetHeading = deviceOffsetPose2.GetHeading();
+      kt_double angleoffset = atan2(deviceOffsetPose2.GetY(), deviceOffsetPose2.GetX());
+      kt_double correctedHeading = math::NormalizeAngle(sPose.GetHeading());
+      Pose2 worldSensorOffset = Pose2(offsetLength * cos(correctedHeading + angleoffset - offsetHeading),
+                                      offsetLength * sin(correctedHeading + angleoffset - offsetHeading),
+                                      offsetHeading);
+
+      return sPose - worldSensorOffset;
+    }
 
   /**
    * Gets the bounding box of this scan
