### 2021-02-07 171f282
add update() when setting corrected pose after optimizing
