### 2021-02-07 6ce85a0
add update() when setting corrected pose after optimizing
