+++ slam_toolbox/lib/karto_sdk/src/Mapper.cpp
@@ -1990,7 +1990,7 @@ namespace karto
         {
           continue;
         }
-        scan->SetCorrectedPose(iter->second);
+        scan->SetCorrectedPoseAndUpdate(iter->second);
       }
 
       pSolver->Clear();
