### 2021-02-07 e6fa421
add update() when setting corrected pose after optimizing
