+++ lib/karto_sdk/include/karto_sdk/Karto.h
@@ -5480,6 +5480,17 @@ public:
   }
 
   /**
+   * Moves the scan by moving the robot pose to the given location and update point readings.
+   * @param rPose new pose of the robot of this scan
+   */
+  inline void SetCorrectedPoseAndUpdate(const Pose2& rPose)
+  {
+    SetCorrectedPose(rPose);
+
+    Update();
+  }
+
+  /**
    * Gets barycenter of point readings
    */
   inline const Pose2 & GetBarycenterPose() const
