+++ src/slam_mapper.cpp
@@ -85,217 +85,209 @@ karto::Pose2 SMapper::toKartoPose(const tf2::Transform & pose) const
 }
 
 /*****************************************************************************/
-void SMapper::configure(const rclcpp::Node::SharedPtr & node)
+void SMapper::configure(const rclcpp::Node::SharedPtr &node)
 /*****************************************************************************/
 {
   bool use_scan_matching = true;
-  if (!node->has_parameter("use_scan_matching") &&
-    node->declare_parameter("use_scan_matching", use_scan_matching))
-  {
-    mapper_->setParamUseScanMatching(use_scan_matching);
+  if (!node->has_parameter("use_scan_matching")) {
+    node->declare_parameter("use_scan_matching", use_scan_matching);
   }
+  node->get_parameter("use_scan_matching", use_scan_matching);
+  mapper_->setParamUseScanMatching(use_scan_matching);
 
   bool use_scan_barycenter = true;
-  if (!node->has_parameter("use_scan_barycenter") &&
-    node->declare_parameter("use_scan_barycenter", use_scan_barycenter))
-  {
-    mapper_->setParamUseScanBarycenter(use_scan_barycenter);
+  if (!node->has_parameter("use_scan_barycenter")) {
+    node->declare_parameter("use_scan_barycenter", use_scan_barycenter);
   }
+  node->get_parameter("use_scan_barycenter", use_scan_barycenter);
+  mapper_->setParamUseScanBarycenter(use_scan_barycenter);
 
   double minimum_travel_distance = 0.5;
-  if (!node->has_parameter("minimum_travel_distance") &&
-    node->declare_parameter("minimum_travel_distance",
-    minimum_travel_distance))
-  {
-    mapper_->setParamMinimumTravelDistance(minimum_travel_distance);
+  if (!node->has_parameter("minimum_travel_distance")) {
+    node->declare_parameter("minimum_travel_distance", minimum_travel_distance);
   }
+  node->get_parameter("minimum_travel_distance", minimum_travel_distance);
+  mapper_->setParamMinimumTravelDistance(minimum_travel_distance);
 
   double minimum_travel_heading = 0.5;
-  if (!node->has_parameter("minimum_travel_heading") &&
-    node->declare_parameter("minimum_travel_heading",
-    minimum_travel_heading))
-  {
-    mapper_->setParamMinimumTravelHeading(minimum_travel_heading);
+  if (!node->has_parameter("minimum_travel_heading")) {
+    node->declare_parameter("minimum_travel_heading", minimum_travel_heading);
   }
+  node->get_parameter("minimum_travel_heading", minimum_travel_heading);
+  mapper_->setParamMinimumTravelHeading(minimum_travel_heading);
 
   int scan_buffer_size = 10;
-  if (!node->has_parameter("scan_buffer_size") &&
-    node->declare_parameter("scan_buffer_size", scan_buffer_size))
-  {
-    mapper_->setParamScanBufferSize(scan_buffer_size);
+  if (!node->has_parameter("scan_buffer_size")) {
+    node->declare_parameter("scan_buffer_size", scan_buffer_size);
   }
+  node->get_parameter("scan_buffer_size", scan_buffer_size);
+  mapper_->setParamScanBufferSize(scan_buffer_size);
+
 
   double scan_buffer_maximum_scan_distance = 10;
-  if (!node->has_parameter("scan_buffer_maximum_scan_distance") &&
-    node->declare_parameter("scan_buffer_maximum_scan_distance",
-    scan_buffer_maximum_scan_distance))
-  {
-    mapper_->setParamScanBufferMaximumScanDistance(scan_buffer_maximum_scan_distance);
+  if (!node->has_parameter("scan_buffer_maximum_scan_distance")) {
[…]
