### 2021-03-02 89c4d60
Initial pose reset melodic (#335)
* first skeleton for clearing observation buffer of localization measurements

* clearing last and running scans for reset of state in localization mode
