+++ slam_toolbox/lib/karto_sdk/include/karto_sdk/Mapper.h
@@ -1585,6 +1585,18 @@ namespace karto
     void SetLastScan(LocalizedRangeScan* pScan);
 
     /**
+     * Clears the laser scan of device
+     * @param pScan
+     */
+    void ClearLastScan(LocalizedRangeScan* pScan);
+
+    /**
+     * Clears the laser scan of device name
+     * @param pScan
+     */
+    void ClearLastScan(const Name& name);
+
+    /**
      * Gets the scan with the given unique id
      * @param id
      * @return scan
@@ -1950,6 +1962,7 @@ namespace karto
     kt_bool ProcessAgainstNodesNearBy(LocalizedRangeScan* pScan);
     kt_bool ProcessLocalization(LocalizedRangeScan* pScan);
     kt_bool RemoveNodeFromGraph(Vertex<LocalizedRangeScan>*);
+    void ClearLocalizationBuffer();
 
     /**
      * Returns all processed scans added to the mapper.
