### 2021-03-02 709e11e
Initial pose reset noetic (#337)
* first skeleton for clearing observation buffer of localization measurements

* clearing last and running scans for reset of state in localization mode
