+++ slam_toolbox/src/slam_toolbox_localization.cpp
@@ -29,6 +29,9 @@ LocalizationSlamToolbox::LocalizationSlamToolbox(ros::NodeHandle& nh)
   processor_type_ = PROCESS_LOCALIZATION;
   localization_pose_sub_ = nh.subscribe("/initialpose", 1,
     &LocalizationSlamToolbox::localizePoseCallback, this);
+  clear_localization_ = nh.advertiseService(
+    "clear_localization_buffer",
+    &LocalizationSlamToolbox::clearLocalizationBuffer, this);
 
   std::string filename;
   geometry_msgs::Pose2D pose;
@@ -56,6 +59,17 @@ LocalizationSlamToolbox::LocalizationSlamToolbox(ros::NodeHandle& nh)
   // in localization mode, disable map saver
   map_saver_.reset();
   return;
+}
+
+/*****************************************************************************/
+bool LocalizationSlamToolbox::clearLocalizationBuffer(
+  std_srvs::Empty::Request& req,
+  std_srvs::Empty::Response& resp)
+/*****************************************************************************/
+{
+  ROS_INFO("LocalizationSlamToolbox: Clearing localization buffer.");
+  smapper_->clearLocalizationBuffer();
+  return true;
 }
 
 /*****************************************************************************/
@@ -204,6 +218,8 @@ void LocalizationSlamToolbox::localizePoseCallback(const
   }
 
   first_measurement_ = true;
+
+  smapper_->clearLocalizationBuffer();
 
   ROS_INFO("LocalizePoseCallback: Localizing to: (%0.2f %0.2f), theta=%0.2f",
     msg->pose.pose.position.x, msg->pose.pose.position.y,
