### 2021-03-02 ae1bdd3
adding reset can buffer on relocalization update
