+++ lib/karto_sdk/include/karto_sdk/Mapper.h
@@ -1604,6 +1604,18 @@ public:
   void SetLastScan(LocalizedRangeScan * pScan);
 
   /**
+   * Clears the laser scan of device
+   * @param pScan
+   */
+  void ClearLastScan(LocalizedRangeScan* pScan);
+
+  /**
+   * Clears the laser scan of device name
+   * @param pScan
+   */
+  void ClearLastScan(const Name& name);
+
+  /**
    * Gets the scan with the given unique id
    * @param id
    * @return scan
@@ -1965,6 +1977,7 @@ public:
   kt_bool ProcessAgainstNodesNearBy(LocalizedRangeScan * pScan);
   kt_bool ProcessLocalization(LocalizedRangeScan * pScan);
   kt_bool RemoveNodeFromGraph(Vertex<LocalizedRangeScan> *);
+  void ClearLocalizationBuffer();
 
   /**
    * Returns all processed scans added to the mapper.
