### 2021-03-02 912a58c
adding reset can buffer on relocalization update (#338)
