### 2021-03-02 12eb539
adding reset can buffer on relocalization update (#340)
