+++ include/slam_toolbox/slam_toolbox_localization.hpp
@@ -21,6 +21,7 @@
 
 #include <memory>
 #include "slam_toolbox/slam_toolbox_common.hpp"
+#include "std_srvs/srv/empty.hpp"
 
 namespace slam_toolbox
 {
@@ -37,6 +38,10 @@ protected:
     sensor_msgs::msg::LaserScan::ConstSharedPtr scan) override;
   void localizePoseCallback(
     const geometry_msgs::msg::PoseWithCovarianceStamped::SharedPtr msg);
+  bool clearLocalizationBuffer(
+    const std::shared_ptr<rmw_request_id_t> request_header,
+    const std::shared_ptr<std_srvs::srv::Empty::Request> req,
+    std::shared_ptr<std_srvs::srv::Empty::Response> resp);
 
   bool serializePoseGraphCallback(
     const std::shared_ptr<rmw_request_id_t> request_header,
@@ -54,6 +59,7 @@ protected:
 
   std::shared_ptr<rclcpp::Subscription<geometry_msgs::msg::PoseWithCovarianceStamped>>
   localization_pose_sub_;
+  std::shared_ptr<rclcpp::Service<std_srvs::srv::Empty> > clear_localization_;
 };
 
 }  // namespace slam_toolbox
