+++ src/slam_mapper.cpp
@@ -53,6 +53,13 @@ void SMapper::setMapper(karto::Mapper * mapper)
 }
 
 /*****************************************************************************/
+void SMapper::clearLocalizationBuffer()
+/*****************************************************************************/
+{
+  mapper_->ClearLocalizationBuffer();
+}
+
+/*****************************************************************************/
 karto::OccupancyGrid * SMapper::getOccupancyGrid(const double & resolution)
 /*****************************************************************************/
 {
