+++ include/slam_toolbox/slam_mapper.hpp
@@ -56,6 +56,8 @@ public:
   void setMapper(karto::Mapper * mapper);
   karto::Mapper * getMapper();
 
+  void clearLocalizationBuffer();
+
 protected:
   std::unique_ptr<karto::Mapper> mapper_;
 };
