### 2021-03-02 e104453
adding scan reset in localization mode
