### 2021-03-02 092e5e8
adding scan reset in localization mode (#341)
