### 2021-03-02 75179f3
fix initial scan not added to localization buffer (#333)
