+++ lib/karto_sdk/src/Mapper.cpp
@@ -177,7 +177,7 @@ namespace karto
         squaredDistance = frontScanPose.GetPosition().SquaredDistance(backScanPose.GetPosition());
       }
     }
-    
+
     /**
      * Finds and replaces a scan from m_scans with NULL
      * @param pScan
@@ -230,7 +230,7 @@ namespace karto
     LocalizedRangeScanMap m_Scans;
     LocalizedRangeScanVector m_RunningScans;
     LocalizedRangeScan* m_pLastScan;
-    kt_int32u m_NextStateId;    
+    kt_int32u m_NextStateId;
 
     kt_int32u m_RunningBufferMaximumSize;
     kt_double m_RunningBufferMaximumDistance;
@@ -341,7 +341,7 @@ namespace karto
   void MapperSensorManager::RemoveScan(LocalizedRangeScan* pScan)
   {
     GetScanManager(pScan)->RemoveScan(pScan);
-    
+
     LocalizedRangeScanMap::iterator it = m_Scans.find(pScan->GetStateId());
     if (it != m_Scans.end())
     {
@@ -2723,7 +2723,7 @@ namespace karto
 	  return false;
   }
 
-  kt_bool Mapper::ProcessAgainstNodesNearBy(LocalizedRangeScan* pScan)
+  kt_bool Mapper::ProcessAgainstNodesNearBy(LocalizedRangeScan* pScan, kt_bool localizationMode)
   {
     if (pScan != NULL)
     {
@@ -2773,10 +2773,11 @@ namespace karto
       // add scan to buffer and assign id
       m_pMapperSensorManager->AddScan(pScan);
 
+      Vertex<LocalizedRangeScan> * scan_vertex = NULL;
       if (m_pUseScanMatching->GetValue())
       {
         // add to graph
-        m_pGraph->AddVertex(pScan);
+        scan_vertex =m_pGraph->AddVertex(pScan);
         m_pGraph->AddEdges(pScan, covariance);
 
         m_pMapperSensorManager->AddRunningScan(pScan);
@@ -2793,6 +2794,10 @@ namespace karto
       }
 
       m_pMapperSensorManager->SetLastScan(pScan);
+
+      if (localizationMode){
+      AddScanToLocalizationBuffer(pScan, scan_vertex);
+      }
 
       return true;
     }
@@ -2835,7 +2840,7 @@ namespace karto
         pScan->GetOdometricPose()));
     }
 
-    // test if scan is outside minimum boundary 
+    // test if scan is outside minimum boundary
     // or if heading is larger then minimum heading
     if (!HasMovedEnough(pScan, pLastScan))
     {
@@ -2867,7 +2872,7 @@ namespace karto
       m_pGraph->AddEdges(pScan, covariance);
 
       m_pMapperSensorManager->AddRunningScan(pScan);
-      
+
       if (m_pDoLoopClosing->GetValue())
       {
         std::vector<Name> deviceNames = m_pMapperSensorManager->GetSensorNames();
@@ -2879,7 +2884,13 @@ namespace karto
     }
 
     m_pMapperSensorManager->SetLastScan(pScan);
+    AddScanToLocalizationBuffer(pScan, scan_vertex);
 
+    return true;
+  }
+
+  void Mapper::AddScanToLocalizationBuffer(LocalizedRangeScan *pScan, Vertex <LocalizedRangeScan> *scan_vertex)
+  {
     // generate the info to store and later decay, outside of dataset
     if (m_LocalizationScanVertices.size() > getParamScanBufferSize())
     {
@@ -2904,8 +2915,6 @@ namespace karto
     lsv.scan = pScan;
     lsv.vertex = scan_vertex;
[…]
