### 2021-03-02 3911bc8
fix initial scan not added to localization buffer (#333)
