### 2021-03-02 33c7ef0
check localization buffer size after adding new scan
(fixes buffer size off-by-one error)
