### 2021-03-02 cabc770
check localization buffer size after adding new scan
(fixes buffer size off-by-one error)
