### 2021-03-02 437d64c
fix the issue (#331) that matcher parameters were ignored [foxy] (#332)
* fixed matcher parameter handling #331

* fixed formating

Co-authored-by: Steve Macenski <stevenmacenski@gmail.com>

Co-authored-by: Steve Macenski <stevenmacenski@gmail.com>
