### 2021-03-03 3a3800b
Cell value is assigned to hitRatio scaled
