### 2021-03-03 9ee627f
Output map is now the scaled raw value of hitRatio
