### 2021-03-03 5fbb151
Add LaunchConfiguration params_file
Signed-off-by: Victor Lopez <victor.lopez@pal-robotics.com>
