+++ src/merge_maps_kinematic.cpp
@@ -51,9 +51,9 @@ void MergeMapsKinematic::configure()
 
   tfB_ = std::make_unique<tf2_ros::TransformBroadcaster>(shared_from_this());
 
-  // interactive_server_ =
-  //   std::make_unique<interactive_markers::InteractiveMarkerServer>(
-  //   "merge_maps_tool","",true);
+  interactive_server_ =
+    std::make_unique<interactive_markers::InteractiveMarkerServer>(
+    "merge_maps_tool", shared_from_this());
 }
 
 /*****************************************************************************/
@@ -146,12 +146,12 @@ bool MergeMapsKinematic::addSubmapCallback(
   m.pose.position.y = transform.getOrigin().getY();
   m.id = num_submaps_;
 
-  // TODO(stevemacenski): waiting on interactive markers to be ported
-  // visualization_msgs::InteractiveMarker int_marker =
-  //   vis_utils::toInteractiveMarker(m, 2.4);
-  // interactive_server_->insert(int_marker,
-  //   boost::bind(&MergeMapsKinematic::processInteractiveFeedback, this, _1));
-  // interactive_server_->applyChanges();
+  visualization_msgs::msg::InteractiveMarker int_marker =
+    vis_utils::toInteractiveMarker(m, 2.4, shared_from_this());
+  interactive_server_->insert(
+    int_marker,
+    std::bind(&MergeMapsKinematic::processInteractiveFeedback, this, std::placeholders::_1));
+  interactive_server_->applyChanges();
 
   RCLCPP_INFO(get_logger(),
     "Map %s was loaded into topic %s!", req->filename.c_str(),
@@ -304,8 +304,7 @@ void MergeMapsKinematic::kartoToROSOccupancyGrid(
 
 /*****************************************************************************/
 void MergeMapsKinematic::processInteractiveFeedback(
-  const
-  visualization_msgs::msg::InteractiveMarkerFeedback::SharedPtr feedback)
+  visualization_msgs::msg::InteractiveMarkerFeedback::ConstSharedPtr feedback)
 /*****************************************************************************/
 {
   const int id = std::stoi(feedback->marker_name, nullptr, 10);
