+++ src/loop_closure_assistant.cpp
@@ -24,8 +24,6 @@
 namespace loop_closure_assistant
 {
 
-// TODO(stevemacenski): Need interactive markers ported to ROS2
-
 /*****************************************************************************/
 LoopClosureAssistant::LoopClosureAssistant(
   rclcpp::Node::SharedPtr node,
@@ -33,105 +31,120 @@ LoopClosureAssistant::LoopClosureAssistant(
   laser_utils::ScanHolder * scan_holder,
   PausedState & state, ProcessType & processor_type)
 : mapper_(mapper), scan_holder_(scan_holder),
-  /*interactive_mode_(false),*/ node_(node), state_(state),
+  interactive_mode_(false), node_(node), state_(state),
   processor_type_(processor_type)
 /*****************************************************************************/
 {
   node_->declare_parameter("paused_processing", false);
   node_->set_parameter(rclcpp::Parameter("paused_processing", false));
+  node_->declare_parameter("interactive_mode", false);
+  node_->set_parameter(rclcpp::Parameter("interactive_mode", false));
+  node_->declare_parameter("enable_interactive_mode", false);
+  node_->get_parameter("enable_interactive_mode", enable_interactive_mode_);
+
   tfB_ = std::make_unique<tf2_ros::TransformBroadcaster>(node_);
   solver_ = mapper_->getScanSolver();
 
-  // ssClear_manual_ = node.advertiseService("clear_changes",
-  //   &LoopClosureAssistant::clearChangesCallback, this);
-  // ssLoopClosure_ = node.advertiseService("manual_loop_closure",
-  //   &LoopClosureAssistant::manualLoopClosureCallback, this);
-  // scan_publisher_ = node.advertise<sensor_msgs::LaserScan>(
-  //   "scan_visualization",10);
-  // interactive_server_ =
-  //   std::make_unique<interactive_markers::InteractiveMarkerServer>(
-  //   "slam_toolbox","",true);
-  // ssInteractive_ = node.advertiseService("toggle_interactive_mode",
-  //   &LoopClosureAssistant::interactiveModeCallback,this);
-  // node.setParam("interactive_mode", interactive_mode_);
-  // node.param("enable_interactive_mode", enable_interactive_mode_, false);
+  ssClear_manual_ = node_->create_service<slam_toolbox::srv::Clear>
+      ("clear_changes", std::bind(&LoopClosureAssistant::clearChangesCallback, 
+      this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
+  
+  ssLoopClosure_ = node_->create_service<slam_toolbox::srv::LoopClosure>
+      ("manual_loop_closure", std::bind(&LoopClosureAssistant::manualLoopClosureCallback,
+      this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
+  
+  scan_publisher_ = node_->create_publisher<sensor_msgs::msg::LaserScan>(
+    "scan_visualization",10);
+  interactive_server_ = std::make_unique<interactive_markers::InteractiveMarkerServer>(
+    "slam_toolbox",
+    node_->get_node_base_interface(),
+    node_->get_node_clock_interface(),
+    node_->get_node_logging_interface(),
+    node_->get_node_topics_interface(),
+    node_->get_node_services_interface());
+  ssInteractive_ = node_->create_service<slam_toolbox::srv::ToggleInteractive>
+      ("toggle_interactive_mode", std::bind(&LoopClosureAssistant::interactiveModeCallback,
+      this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
+
+
   marker_publisher_ = node_->create_publisher<visualization_msgs::msg::MarkerArray>(
     "graph_visualization", rclcpp::QoS(1));
[…]

[…]
