### 2021-03-04 50f0fda
clean up cmake warnings
