+++ src/loop_closure_assistant.cpp
@@ -44,13 +44,13 @@ LoopClosureAssistant::LoopClosureAssistant(
   tfB_ = std::make_unique<tf2_ros::TransformBroadcaster>(node_);
   solver_ = mapper_->getScanSolver();
 
-  ssClear_manual_ = node_->create_service<slam_toolbox::srv::Clear>
-      ("/slam_toolbox/clear_changes", std::bind(&LoopClosureAssistant::clearChangesCallback, 
-      this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
+  ssClear_manual_ = node_->create_service<slam_toolbox::srv::Clear>(
+    "/slam_toolbox/clear_changes", std::bind(&LoopClosureAssistant::clearChangesCallback, 
+    this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
   
-  ssLoopClosure_ = node_->create_service<slam_toolbox::srv::LoopClosure>
-      ("/slam_toolbox/manual_loop_closure", std::bind(&LoopClosureAssistant::manualLoopClosureCallback,
-      this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
+  ssLoopClosure_ = node_->create_service<slam_toolbox::srv::LoopClosure>(
+    "/slam_toolbox/manual_loop_closure", std::bind(&LoopClosureAssistant::manualLoopClosureCallback,
+    this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
   
   scan_publisher_ = node_->create_publisher<sensor_msgs::msg::LaserScan>(
     "/slam_toolbox/scan_visualization",10);
@@ -61,9 +61,9 @@ LoopClosureAssistant::LoopClosureAssistant(
     node_->get_node_logging_interface(),
     node_->get_node_topics_interface(),
     node_->get_node_services_interface());
-  ssInteractive_ = node_->create_service<slam_toolbox::srv::ToggleInteractive>
-      ("/slam_toolbox/toggle_interactive_mode", std::bind(&LoopClosureAssistant::interactiveModeCallback,
-      this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
+  ssInteractive_ = node_->create_service<slam_toolbox::srv::ToggleInteractive>(
+    "/slam_toolbox/toggle_interactive_mode", std::bind(&LoopClosureAssistant::interactiveModeCallback,
+    this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3));
 
 
   marker_publisher_ = node_->create_publisher<visualization_msgs::msg::MarkerArray>(
