### 2021-03-04 a4cd5a2
Ros2 interactive markers and rviz port (#349)
* initial port work

* loading plugin

* clean up cmake warnings

* correcting namespacing for ROS21

* some last cleanup
