+++ rviz_plugin/slam_toolbox_rviz_plugin.hpp
@@ -40,7 +40,7 @@
 #include "rclcpp/rclcpp.hpp"
 #include "rviz_common/panel.hpp"
 #include "slam_toolbox/toolbox_msgs.hpp"
-#include "geometry_msgs/msg/PoseWithCovarianceStamped.hpp"
+#include "geometry_msgs/msg/pose_with_covariance_stamped.hpp"
 
 
 class QLineEdit;
@@ -68,12 +68,11 @@ class SlamToolboxPlugin : public rviz_common::Panel
   Q_OBJECT
 
 public:
-  explicit SlamToolboxPlugin(QWidget * parent);
-  SlamToolboxPlugin() {SlamToolboxPlugin(nullptr);}
+  explicit SlamToolboxPlugin(QWidget * parent = 0);
 
-  ~SlamToolboxPlugin();
+  virtual ~SlamToolboxPlugin();
 
-public Q_SLOTS:
+private Q_SLOTS:
   void ClearChanges();
   void SaveChanges();
   void SaveMap();
@@ -157,7 +156,7 @@ protected:
   
   rclcpp::Subscription<geometry_msgs::msg::PoseWithCovarianceStamped>::SharedPtr _initialposeSub;
   
-  void InitialPoseCallback(const geometry_msgs::msg::PoseWithCovarianceStamped::SharedPtr& pose);
+  void InitialPoseCallback(const geometry_msgs::msg::PoseWithCovarianceStamped::SharedPtr pose);
 };
 
 }  // namespace slam_toolbox
