### 2021-03-04 7eb866f
Ros2 interactive markers and rviz port (#349) (#351)
* initial port work

* loading plugin

* clean up cmake warnings

* correcting namespacing for ROS21

* some last cleanup
