+++ rviz_plugin/slam_toolbox_rviz_plugin.cpp
@@ -17,9 +17,10 @@
 /* Author: Steven Macenski */
 
 // Header
-#include "rviz_plugin/slam_toolbox_rviz_plugin.h"
+#include "rviz_plugin/slam_toolbox_rviz_plugin.hpp"
 // ROS
 #include <tf2_ros/transform_listener.h>
+#include <tf2/convert.h>
 // QT
 #include <QPushButton>
 #include <QCheckBox>
@@ -51,8 +52,10 @@ SlamToolboxPlugin::SlamToolboxPlugin(QWidget * parent)
   interactive = ros_node_->declare_parameter(
     "/slam_toolbox/interactive_mode", interactive);
     
-  _initialposeSub = ros_node_->create_subscription<geometry_msgs::msg::PoseWithCovarianceStamped>(
-    "/initialpose", 10, std::bind(&SlamToolboxPlugin::InitialPoseCallback, this, _1));
+  _initialposeSub =
+    ros_node_->create_subscription<geometry_msgs::msg::PoseWithCovarianceStamped>(
+    "/initialpose", 10,
+    std::bind(&SlamToolboxPlugin::InitialPoseCallback, this, std::placeholders::_1));
 
   _serialize =
     ros_node_->create_client<slam_toolbox::srv::SerializePoseGraph>(
@@ -251,11 +254,14 @@ SlamToolboxPlugin::~SlamToolboxPlugin()
 }
   
 /*****************************************************************************/
-void SlamToolboxPlugin::InitialPoseCallback(const geometry_msgs::msg::PoseWithCovarianceStamped::SharedPtr& msg)
+void SlamToolboxPlugin::InitialPoseCallback(
+  const geometry_msgs::msg::PoseWithCovarianceStamped::SharedPtr msg)
 /*****************************************************************************/
 {
   _match_type = PROCESS_NEAR_REGION_CMT;
-  ROS_INFO("Setting initial pose from rviz; you can now deserialize a map given that pose.");
+  RCLCPP_INFO(
+    ros_node_->get_logger(),
+    "Setting initial pose from rviz; you can now deserialize a map given that pose.");
   _radio2->setChecked(true);
   _line5->setText(QString::number(msg->pose.pose.position.x, 'f', 2));
   _line6->setText(QString::number(msg->pose.pose.position.y, 'f', 2));
@@ -278,7 +284,7 @@ void SlamToolboxPlugin::SerializeMap()
 
   if (rclcpp::spin_until_future_complete(ros_node_, result_future,
     std::chrono::seconds(5)) !=
-    rclcpp::executor::FutureReturnCode::SUCCESS)
+    rclcpp::FutureReturnCode::SUCCESS)
   {
     RCLCPP_WARN(ros_node_->get_logger(),
       "SlamToolbox: Failed to serialize"
@@ -307,7 +313,7 @@ void SlamToolboxPlugin::DeserializeMap()
     }
     catch (const std::invalid_argument& ia)
     {
-      ROS_WARN("Initial pose invalid.");
+      RCLCPP_WARN(ros_node_->get_logger(), "Initial pose invalid.");
       return;
     }
   } else if (_match_type == LOCALIZE_CMT) {
@@ -320,11 +326,12 @@ void SlamToolboxPlugin::DeserializeMap()
     }
     catch (const std::invalid_argument& ia)
     {
-      ROS_WARN("Initial pose invalid.");
+      RCLCPP_WARN(ros_node_->get_logger(), "Initial pose invalid.");
       return;
     }
   } else {
-    RCLCPP_WARN(ros_node_->get_logger(),
+    RCLCPP_WARN(
[…]
