### 2021-03-05 8a0348d
fix declares for multiple sensors
