### 2021-03-05 e4ef463
fix deserialization crash due to invalid sensor ID
