### 2021-03-05 20b6967
fixed formating and naming
requested changes from the code review

Co-authored-by: Steve Macenski <stevenmacenski@gmail.com>
