### 2021-03-06 36e16b3
fix deserialization crash due to invalid sensor ID (#354)
