### 2021-03-06 6bfd26f
fixing deserialization crash due to invalid sensor ID [foxy] (#353)
* fixing deserialization crash due to invalid sensor ID

* fix declares for multiple sensors

* fix typo
