### 2021-03-06 fcbc963
fixing deserialization crash due to invalid sensor IDs (#352)
