### 2021-03-09 997757c
fix initial scan not added to localization buffer (#333)
