### 2021-03-09 1950c1a
fix initial scan not added to localization buffer (#333)
