### 2021-03-09 d98031e
check localization buffer size after adding new scan
(fixes buffer size off-by-one error)
