### 2021-03-09 7720e30
check localization buffer size after adding new scan
(fixes buffer size off-by-one error)
