### 2021-03-09 #2783 @mosteo
I am finished, sorry for the delay. A couple of observations:

In the SLAM tutorial at https://navigation.ros.org/tutorials/docs/navigation2_with_slam.html, the command at step 1:

` ros2 launch nav2_bringup nav2_navigation_launch.py` should be
`ros2 launch nav2_bringup navigation_launch.py`

At step 4 with the shortcut, the last `True` is given in lowercase and should be in uppercase: `ros2 launch nav2_bringup tb3_simulation_launch.py slam:=true`.

I could not load the plugin for RViz mentioned in the readme in RViz2. Is this ROS1 only? Is it not bundled with Foxy packages? Am I missing something? If not the latter, this should be clarified in the readme.
