### 2021-03-09 877784a
fix issue (#333) initial scan not added to localization buffer [foxy] (#343)
* fix initial scan not added to localization buffer (#333)

* check localization buffer size after adding new scan

(fixes buffer size off-by-one error)

* fixed formating and naming

requested changes from the code review

Co-authored-by: Steve Macenski <stevenmacenski@gmail.com>

Co-authored-by: Steve Macenski <stevenmacenski@gmail.com>
