### 2021-03-09 f467248
fix issue (#333) initial scan not added to localization buffer [meldoic] (#359)
* fix initial scan not added to localization buffer (#333)

* check localization buffer size after adding new scan

(fixes buffer size off-by-one error)
