### 2021-03-10 #2783 @mosteo
> They are bundled and should work, can you explain what you mean by didn't load?

Yes, the navigation panel does show. What I'm missing there is the options about map manipulation mentioned in the README, i.e. the panel shown at the end of section https://github.com/SteveMacenski/slam_toolbox#rviz-plugin. I had tried the `Add Panels -> Add New Panel`, where the same Nav2 panel exists, but nothing else about slam.

Now, I've found that there's a config file `foxy/share/slam_toolbox/config/slam_toolbox_default.rviz`, but that says (in an error dialog):

![image](https://user-images.githubusercontent.com/264377/110607248-4d2a7f00-818b-11eb-9038-245d592b3652.png)

Afterwards there are other errors in the panels and rviz quickly dies.
