### 2021-03-10 #2783 @SteveMacenski
The map manipulation tools are only available in ROS1 due to the incomplete port of interactive markers in ROS2 (this document is talking about ROS1) You can enter interactive mode and see the interactive markers but unfortunately the "dragging" of them to reposition them for manipulation is still unavailable.

That's not an excuse for the rviz config issue, I've removed the slam toolbox rviz config, all it was was the default config file with the SLAM Toolbox plugin pre-loaded. However, rviz2 has renamed a bunch of the namespaces in its port after this was ported. I removed the file to remove this confusion (and any further issues with rviz changes impacting the repo since really it was just preloading a simple panel).
