### 2021-03-10 #2783 @mosteo
Thanks, @SteveMacenski. I'm not that worried about some things being only available for ROS1 (I known the migration to ROS2 is not being trivial), but about the confusion a newcomer may experience.
