### 2021-03-10 #2783 @SteveMacenski
Totally fair - I removed that config file from `ros2` branch and very shortly `foxy` too. Let me know if there's anything more I can do to help!
