### 2021-03-12 f5e08a0
Rename params_file to slam_params_file to avoid name conflicts
See https://github.com/ros-planning/navigation2/issues/2240
