### 2021-03-15 5d397c7
Revert "Rename params_file to slam_params_file to avoid name conflicts (#364)"
This reverts commit 33ad9d431c93f113b753e602c045a64e7a936f40.
