### 2021-03-17 0df5e79
Check params_file arg, and use default if it has no slam_toolbox params
This is a workaround to
https://github.com/ros-planning/navigation2/pull/2243 to keep foxy API
stable by not changing the params_file name.
