### 2021-03-17 58b454c
Use has_node_params in nav2_common
