### 2021-03-18 99e12e5
Check params_file arg, and use default if it has no slam_toolbox params (#368)
* Check params_file arg, and use default if it has no slam_toolbox params

This is a workaround to
https://github.com/ros-planning/navigation2/pull/2243 to keep foxy API
stable by not changing the params_file name.

* Use has_node_params in nav2_common
