### 2021-03-23 cf99c83
fix the issue (#198) that loop closures can rotate the map 180° if the lidar is mounted backwards (#326)
* fix the issue (#198) that loop closures can rotate the map 180° if the lidar is mounted backwards

Set the edges (constraints) based on the robot base frame (base_footprint) instead of sensor frame (laser).

* linting things

* add function GetCorrectedAt to simplify getting robot pose

* add update() when setting corrected pose after optimizing
