### 2021-03-24 SteveMacenski/slam_toolbox#320 @carlosjoserg
Hi @SteveMacenski , apologies for the big delay.

I (re-)read the paper some days ago, reviewed your replies to my issues, and ticked the related check-boxes in the main thread.

The remaining part is about the actual funcionality (and its documentation), which I'd like to go through in the next few days. I'll keep the issue open  just in case.

And thanks for your patience!
