### 2021-03-24 SteveMacenski/slam_toolbox#320 @SteveMacenski
Thanks! By the way, for the still-unmarked box API documentation: http://docs.ros.org/en/noetic/api/slam_toolbox/html/namespaces.html (and https://vimeo.com/378682207 video where I explain all the features as well since that box was a little ambiguous to what it meant. The readme here is also very extensive)

If you have any questions on testing the features, don't hesitate to ask!
