### 2021-03-29 d44f801
handle initialization for deserialized files (#355)
