### 2021-03-29 ce01f6a
handle initialization for deserialized files (#355)
