+++ lib/karto_sdk/src/Mapper.cpp
@@ -158,6 +158,24 @@ public:
   }
 
   /**
+   * Sets running scan buffer size
+   * @param rScanBufferSize
+   */
+  void SetRunningScanBufferSize(const kt_int32u & rScanBufferSize)
+  {
+    m_RunningBufferMaximumSize = rScanBufferSize;
+  }
+
+  /**
+   * Sets running scan buffer maximum distance
+   * @param rScanBufferMaxDistance
+   */
+  void SetRunningScanBufferMaximumDistance(const kt_int32u & rScanBufferMaxDistance)
+  {
+    m_RunningBufferMaximumDistance = rScanBufferMaxDistance;
+  }
+
+  /**
    * Adds scan to vector of running scans
    * @param pScan
    */
@@ -247,8 +265,9 @@ private:
 void MapperSensorManager::RegisterSensor(const Name & rSensorName)
 {
   if (GetScanManager(rSensorName) == NULL) {
-    m_ScanManagers[rSensorName] = new ScanManager(m_RunningBufferMaximumSize,
-        m_RunningBufferMaximumDistance);
+    m_ScanManagers[rSensorName] = new ScanManager(
+      m_RunningBufferMaximumSize,
+      m_RunningBufferMaximumDistance);
   }
 }
 
@@ -379,6 +398,26 @@ void MapperSensorManager::ClearRunningScans(const Name & rSensorName)
 inline kt_int32u MapperSensorManager::GetRunningScanBufferSize(const Name & rSensorName)
 {
   return GetScanManager(rSensorName)->GetRunningScanBufferSize();
+}
+
+void MapperSensorManager::SetRunningScanBufferSize(kt_int32u rScanBufferSize)
+{
+  m_RunningBufferMaximumSize = rScanBufferSize;
+
+  std::vector<Name> names = GetSensorNames();
+  for (uint i = 0; i != names.size(); i++) {
+    GetScanManager(names[i])->SetRunningScanBufferSize(rScanBufferSize);
+  }
+}
+
+void MapperSensorManager::SetRunningScanBufferMaximumDistance(kt_double rScanBufferMaxDistance)
+{
+  m_RunningBufferMaximumDistance = rScanBufferMaxDistance;
+
+  std::vector<Name> names = GetSensorNames();
+  for (uint i = 0; i != names.size(); i++) {
+    GetScanManager(names[i])->SetRunningScanBufferMaximumDistance(rScanBufferMaxDistance);
+  }
 }
 
 /**
@@ -1355,9 +1394,9 @@ MapperGraph::MapperGraph(Mapper * pMapper, kt_double rangeThreshold)
 : m_pMapper(pMapper)
 {
   m_pLoopScanMatcher = ScanMatcher::Create(pMapper,
-      m_pMapper->m_pLoopSearchSpaceDimension->GetValue(),
-      m_pMapper->m_pLoopSearchSpaceResolution->GetValue(),
-      m_pMapper->m_pLoopSearchSpaceSmearDeviation->GetValue(), rangeThreshold);
+    m_pMapper->m_pLoopSearchSpaceDimension->GetValue(),
+    m_pMapper->m_pLoopSearchSpaceResolution->GetValue(),
+    m_pMapper->m_pLoopSearchSpaceSmearDeviation->GetValue(), rangeThreshold);
   assert(m_pLoopScanMatcher);
 
   m_pTraversal = new BreadthFirstTraversal<LocalizedRangeScan>(this);
@@ -1989,6 +2028,18 @@ void MapperGraph::CorrectPoses()
   }
 }
 
+void MapperGraph::UpdateLoopScanMatcher(kt_double rangeThreshold)
+{
+  if (m_pLoopScanMatcher) {
+    delete m_pLoopScanMatcher;
+  }
[…]
