### 2021-03-29 aabb16d
handle initialization for deserialized files (#355)
