### 2021-03-30 4fb8a96
fix parallel access to scans while clearing localization buffer
