### 2021-03-31 d53cee2
Fixed error for function RemoveScan
