### 2021-03-31 03703ea
Fixed error for function RemoveScan (#372)
