+++ slam_toolbox/lib/karto_sdk/src/Mapper.cpp
@@ -342,7 +342,7 @@ namespace karto
   {
     GetScanManager(pScan)->RemoveScan(pScan);
     
-    LocalizedRangeScanMap::iterator it = m_Scans.find(pScan->GetStateId());
+    LocalizedRangeScanMap::iterator it = m_Scans.find(pScan->GetUniqueId());
     if (it != m_Scans.end())
     {
       it->second = NULL;
