### 2021-04-06 5ba6d39
removed redundant code (code review)
