### 2021-04-08 d941f99
fixed timestamped ouput not match input
