### 2021-04-08 0458f6d
fixed timestamped ouput not match input
