### 2021-04-08 d42b9e5
Update params for deserialized files [foxy] (#371)
* handle initialization for deserialized files (#355)

* fix parallel access to scans while clearing localization buffer

* removed redundant code (code review)
