### 2021-04-08 5de4f43
Update params for deserialized files [melodic] (#380)
* handle initialization for deserialized files (#355)

* fix parallel access to scans while clearing localization buffer
