### 2021-04-08 2678f19
Update params for deserialized files [ros2] (#379)
* handle initialization for deserialized files (#355)

* fix parallel access to scans while clearing localization buffer

* removed redundant code (code review)
