+++ src/slam_toolbox_localization.cpp
@@ -79,6 +79,7 @@ bool LocalizationSlamToolbox::clearLocalizationBuffer(
   std::shared_ptr<std_srvs::srv::Empty::Response> resp)
 /*****************************************************************************/
 {
+  boost::mutex::scoped_lock lock(smapper_mutex_);
   RCLCPP_INFO(get_logger(),
     "LocalizationSlamToolbox: Clearing localization buffer.");
   smapper_->clearLocalizationBuffer();
@@ -220,6 +221,7 @@ void LocalizationSlamToolbox::localizePoseCallback(
 
   first_measurement_ = true;
 
+  boost::mutex::scoped_lock lock(smapper_mutex_);
   smapper_->clearLocalizationBuffer();
 
   RCLCPP_INFO(get_logger(),
