### 2021-04-08 1e7ff7a
Update params for deserialized files [dashing] (#378)
* handle initialization for deserialized files (#355)

* fix parallel access to scans while clearing localization buffer
