+++ lib/karto_sdk/src/Mapper.cpp
@@ -152,6 +152,24 @@ namespace karto
     }
 
     /**
+     * Sets running scan buffer size
+     * @param rScanBufferSize
+     */
+    void SetRunningScanBufferSize(const kt_int32u & rScanBufferSize)
+    {
+      m_RunningBufferMaximumSize = rScanBufferSize;
+    }
+
+    /**
+     * Sets running scan buffer maximum distance
+     * @param rScanBufferMaxDistance
+     */
+    void SetRunningScanBufferMaximumDistance(const kt_int32u & rScanBufferMaxDistance)
+    {
+      m_RunningBufferMaximumDistance = rScanBufferMaxDistance;
+    }
+
+    /**
      * Adds scan to vector of running scans
      * @param pScan
      */
@@ -383,6 +401,26 @@ namespace karto
   inline kt_int32u MapperSensorManager::GetRunningScanBufferSize(const Name& rSensorName)
   {
     return GetScanManager(rSensorName)->GetRunningScanBufferSize();
+  }
+
+  void MapperSensorManager::SetRunningScanBufferSize(kt_int32u rScanBufferSize)
+  {
+    m_RunningBufferMaximumSize = rScanBufferSize;
+
+    std::vector<Name> names = GetSensorNames();
+    for (uint i = 0; i != names.size(); i++) {
+      GetScanManager(names[i])->SetRunningScanBufferSize(rScanBufferSize);
+    }
+  }
+
+  void MapperSensorManager::SetRunningScanBufferMaximumDistance(kt_double rScanBufferMaxDistance)
+  {
+    m_RunningBufferMaximumDistance = rScanBufferMaxDistance;
+
+    std::vector<Name> names = GetSensorNames();
+    for (uint i = 0; i != names.size(); i++) {
+      GetScanManager(names[i])->SetRunningScanBufferMaximumDistance(rScanBufferMaxDistance);
+    }
   }
 
   /**
@@ -2022,6 +2060,18 @@ namespace karto
     }
   }
 
+  void MapperGraph::UpdateLoopScanMatcher(kt_double rangeThreshold)
+  {
+    if (m_pLoopScanMatcher) {
+      delete m_pLoopScanMatcher;
+    }
+    m_pLoopScanMatcher = ScanMatcher::Create(m_pMapper,
+      m_pMapper->m_pLoopSearchSpaceDimension->GetValue(),
+      m_pMapper->m_pLoopSearchSpaceResolution->GetValue(),
+      m_pMapper->m_pLoopSearchSpaceSmearDeviation->GetValue(), rangeThreshold);
+    assert(m_pLoopScanMatcher);
+  }
+
   ////////////////////////////////////////////////////////////////////////////////////////
   ////////////////////////////////////////////////////////////////////////////////////////
   ////////////////////////////////////////////////////////////////////////////////////////
@@ -2030,12 +2080,13 @@ namespace karto
    * Default constructor
    */
   Mapper::Mapper()
-    : Module("Mapper")
-    , m_Initialized(false)
-    , m_pSequentialScanMatcher(NULL)
-    , m_pMapperSensorManager(NULL)
-    , m_pGraph(NULL)
-    , m_pScanOptimizer(NULL)
+  : Module("Mapper"),
+    m_Initialized(false),
+    m_Deserialized(false),
+    m_pSequentialScanMatcher(NULL),
+    m_pMapperSensorManager(NULL),
+    m_pGraph(NULL),
+    m_pScanOptimizer(NULL)
   {
     InitializeParameters();
   }
@@ -2043,13 +2094,14 @@ namespace karto
   /**
    * Default constructor
    */
-  Mapper::Mapper(const std::string& rName)
-    : Module(rName)
-    , m_Initialized(false)
-    , m_pSequentialScanMatcher(NULL)
[…]
