### 2021-04-14 199eef1
Return back to from_nanoseconds
