### 2021-04-16 b16275e
cleanup lint violations by ament_cpplint
