+++ src/experimental/slam_toolbox_lifelong.cpp
@@ -85,7 +85,6 @@ void LifelongSlamToolbox::laserCallback(
 {
   // store scan timestamped
   scan_timestamped = scan->header.stamp;
-
   // no odom info
   Pose2 pose;
   if (!pose_helper_->getOdomPose(pose, scan->header.stamp)) {
