+++ src/slam_toolbox_localization.cpp
@@ -120,7 +120,6 @@ void LocalizationSlamToolbox::laserCallback(
 {
   // store scan timestamped
   scan_timestamped = scan->header.stamp;
-
   // no odom info
   Pose2 pose;
   if (!pose_helper_->getOdomPose(pose, scan->header.stamp)) {
