+++ src/slam_toolbox_async.cpp
@@ -37,7 +37,6 @@ void AsynchronousSlamToolbox::laserCallback(
 {
   // store scan timestamped
   scan_timestamped = scan->header.stamp;
-  
   // no odom info
   Pose2 pose;
   if (!pose_helper_->getOdomPose(pose, scan->header.stamp)) {
