+++ src/slam_toolbox_sync.cpp
@@ -67,7 +67,6 @@ void SynchronousSlamToolbox::laserCallback(
 {
   // store scan timestamped
   scan_timestamped = scan->header.stamp;
-
   // no odom info
   Pose2 pose;
   if (!pose_helper_->getOdomPose(pose, scan->header.stamp)) {
