### 2021-04-16 749a3b1
fixed timestamps input&output not synchronized  (#377)
* fixed timestamped ouput not match input

* fixed timestamped ouput not match input

* Return back to from_nanoseconds

* No need for the comment, git history will have this.

* cleanup lint violations by ament_cpplint

Co-authored-by: SirawatS <SirawatS@arv.co.th>
