### 2021-04-19 b3e6381
change karto library name so that the package is installable alongside openkarto (#383)
Signed-off-by: Mikael Arguedas <mikael.arguedas@gmail.com>
