### 2021-04-27 2f9bc2f
fix issue (#180) that localization mode may terminate unexpectedly with exception about pthread_mutex_lock
