### 2021-04-27 86b0fc0
fix issue (#180) that localization mode may terminate unexpectedly with exception about pthread_mutex_lock (#386)
* fix the issue (#198) that loop closures can rotate the map 180° if the lidar is mounted backwards

Set the edges (constraints) based on the robot base frame (base_footprint) instead of sensor frame (laser).

* linting things

* add function GetCorrectedAt to simplify getting robot pose

* add update() when setting corrected pose after optimizing

* fix issue (#180) that localization mode may terminate unexpectedly with exception about pthread_mutex_lock
