### 2021-04-27 a72ad2e
Revert "fix issue (#180) that localization mode may terminate unexpectedly with exception about pthread_mutex_lock (#386)" (#389)
This reverts commit 86b0fc08f4e84e685ebf6592182ed2d87741e9c9.
