### 2021-04-28 4d66dda
set c++17 standard for the whole project, use from_second insteads of incorrect from_nanoseconds
