### 2021-04-28 2f869c0
fix issue (#180) that localization mode may terminate unexpectedly with exception about pthread_mutex_lock (#388)
* fix issue (#180) that localization mode may terminate unexpectedly with exception about pthread_mutex_lock

* set c++17 standard for the whole project, use from_second insteads of incorrect from_nanoseconds
