+++ lib/karto_sdk/include/karto_sdk/Karto.h
@@ -49,6 +49,7 @@
 #include <iomanip>
 #include <sstream>
 #include <stdexcept>
+#include <shared_mutex>
 
 #ifdef USE_POCO
 #include <Poco/Mutex.h>
@@ -5434,7 +5435,7 @@ public:
   }
 
 private:
-  mutable boost::shared_mutex m_Lock;
+  mutable std::shared_mutex m_Lock;
 
 public:
   /**
@@ -5495,11 +5496,11 @@ public:
    */
   inline const Pose2 & GetBarycenterPose() const
   {
-    boost::shared_lock<boost::shared_mutex> lock(m_Lock);
+    std::shared_lock<std::shared_mutex> lock(m_Lock);
     if (m_IsDirty) {
       // throw away constness and do an update!
       lock.unlock();
-      boost::unique_lock<boost::shared_mutex> uniqueLock(m_Lock);
+      std::unique_lock<std::shared_mutex> uniqueLock(m_Lock);
       const_cast<LocalizedRangeScan *>(this)->Update();
     }
 
@@ -5518,11 +5519,11 @@ public:
    */
   inline Pose2 GetReferencePose(kt_bool useBarycenter) const
   {
-    boost::shared_lock<boost::shared_mutex> lock(m_Lock);
+    std::shared_lock<std::shared_mutex> lock(m_Lock);
     if (m_IsDirty) {
       // throw away constness and do an update!
       lock.unlock();
-      boost::unique_lock<boost::shared_mutex> uniqueLock(m_Lock);
+      std::unique_lock<std::shared_mutex> uniqueLock(m_Lock);
       const_cast<LocalizedRangeScan *>(this)->Update();
     }
 
@@ -5589,11 +5590,11 @@ public:
    */
   inline const BoundingBox2 & GetBoundingBox() const
   {
-    boost::shared_lock<boost::shared_mutex> lock(m_Lock);
+    std::shared_lock<std::shared_mutex> lock(m_Lock);
     if (m_IsDirty) {
       // throw away constness and do an update!
       lock.unlock();
-      boost::unique_lock<boost::shared_mutex> uniqueLock(m_Lock);
+      std::unique_lock<std::shared_mutex> uniqueLock(m_Lock);
       const_cast<LocalizedRangeScan *>(this)->Update();
     }
 
@@ -5610,11 +5611,11 @@ public:
    */
   inline const PointVectorDouble & GetPointReadings(kt_bool wantFiltered = false) const
   {
-    boost::shared_lock<boost::shared_mutex> lock(m_Lock);
+    std::shared_lock<std::shared_mutex> lock(m_Lock);
     if (m_IsDirty) {
       // throw away constness and do an update!
       lock.unlock();
-      boost::unique_lock<boost::shared_mutex> uniqueLock(m_Lock);
+      std::unique_lock<std::shared_mutex> uniqueLock(m_Lock);
       const_cast<LocalizedRangeScan *>(this)->Update();
     }
 
