+++ slam_toolbox/lib/karto_sdk/include/karto_sdk/Karto.h
@@ -28,6 +28,7 @@
 #include <iomanip>
 #include <sstream>
 #include <stdexcept>
+#include <shared_mutex>
 
 #include <math.h>
 #include <float.h>
@@ -5528,7 +5529,7 @@ namespace karto
     }
 
   private:
-    mutable boost::shared_mutex m_Lock;
+    mutable std::shared_mutex m_Lock;
 
   public:
     /**
@@ -5589,12 +5590,12 @@ namespace karto
      */
     inline const Pose2& GetBarycenterPose() const
     {
-      boost::shared_lock<boost::shared_mutex> lock(m_Lock);
+      std::shared_lock<std::shared_mutex> lock(m_Lock);
       if (m_IsDirty)
       {
         // throw away constness and do an update!
         lock.unlock();
-        boost::unique_lock<boost::shared_mutex> uniqueLock(m_Lock);
+        std::unique_lock<std::shared_mutex> uniqueLock(m_Lock);
         const_cast<LocalizedRangeScan*>(this)->Update();
       }
 
@@ -5613,12 +5614,12 @@ namespace karto
      */
     inline Pose2 GetReferencePose(kt_bool useBarycenter) const
     {
-      boost::shared_lock<boost::shared_mutex> lock(m_Lock);
+      std::shared_lock<std::shared_mutex> lock(m_Lock);
       if (m_IsDirty)
       {
         // throw away constness and do an update!
         lock.unlock();
-        boost::unique_lock<boost::shared_mutex> uniqueLock(m_Lock);
+        std::unique_lock<std::shared_mutex> uniqueLock(m_Lock);
         const_cast<LocalizedRangeScan*>(this)->Update();
       }
 
@@ -5685,12 +5686,12 @@ namespace karto
      */
     inline const BoundingBox2& GetBoundingBox() const
     {
-      boost::shared_lock<boost::shared_mutex> lock(m_Lock);
+      std::shared_lock<std::shared_mutex> lock(m_Lock);
       if (m_IsDirty)
       {
         // throw away constness and do an update!
         lock.unlock();
-        boost::unique_lock<boost::shared_mutex> uniqueLock(m_Lock);
+        std::unique_lock<std::shared_mutex> uniqueLock(m_Lock);
         const_cast<LocalizedRangeScan*>(this)->Update();
       }
 
@@ -5707,12 +5708,12 @@ namespace karto
      */
     inline const PointVectorDouble& GetPointReadings(kt_bool wantFiltered = false) const
     {
-      boost::shared_lock<boost::shared_mutex> lock(m_Lock);
+      std::shared_lock<std::shared_mutex> lock(m_Lock);
       if (m_IsDirty)
       {
         // throw away constness and do an update!
         lock.unlock();
-        boost::unique_lock<boost::shared_mutex> uniqueLock(m_Lock);
+        std::unique_lock<std::shared_mutex> uniqueLock(m_Lock);
         const_cast<LocalizedRangeScan*>(this)->Update();
       }
 
