### 2021-05-11 SteveMacenski/slam_toolbox#320 @carlosjoserg
Hi @SteveMacenski

Thanks for your patience (and for sharing).

Finally got to test part of the functionality on `melodic` and `noetic`. It all looks good as described, rviz plugin loaded fine too.

I'm marking all items resolved, closing the issue, and wish for a reduction in the amount of parameters (not a particular request for slam_toolbox though, but for anything in general ;)).

Best,
