### 2021-05-11 #2783 @SteveMacenski
@arfon thanks! The newest releases contain all of the change requested and patched as part of the review. I just verified by looking back at all of the review comments/threads and cross referencing change requested with commits in those releases:
- ROS2 2.4.0 https://github.com/SteveMacenski/slam_toolbox/releases/tag/2.4.0
- ROS(1) 1.5.5 https://github.com/SteveMacenski/slam_toolbox/releases/tag/1.5.5

Zenodo archive link (named with the same name and with both authors): https://doi.org/10.5281/zenodo.4749721

10.5281/zenodo.4749721
