### 2021-05-12 #2783 @arfon
@SteveMacenski - a couple of final things:

- Did you check the DOI suggestions from Whedon above? If any of them are correct, please update your BibTeX to include them.
- I made a small change to your paper in https://github.com/SteveMacenski/slam_toolbox/pull/392 to make it read a little more clearly (I think).
