### 2021-05-12 #2783 @SteveMacenski
Hi,

- [Googling the DOIs](https://www.google.com/search?channel=fs&client=ubuntu&q=10.1109%2Firos45743.2020.9341207) all gave me the papers that are also matching the titles. I'm not exactly sure what the issue is there, but they seem to all be valid from my checking
- Merged
