### 2021-05-13 #2783 @arfon
> Googling the DOIs all gave me the papers that are also matching the titles. I'm not exactly sure what the issue is there, but they seem to all be valid from my checking

Great! Then please update your `paper.bib` to include them. For example, for `The Marathon 2: A Navigation System`, if the suggested DOI from Whedon is correct then you need to add `doi=10.1109/iros45743.2020.9341207` to [this part](https://github.com/SteveMacenski/slam_toolbox/blob/joss/paper.bib#L48-L54) of your BibTeX file (we need you to do this for all of the potential missing DOIs please).
