+++ paper.bib
@@ -51,20 +51,23 @@
 	year = {2020},
 	booktitle = {2020 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
 	url = {https://github.com/ros-planning/navigation2},
-	pdf = {https://arxiv.org/abs/2003.00368}
+	pdf = {https://arxiv.org/abs/2003.00368},
+	doi = {10.1109/iros45743.2020.9341207}
 }
 @InProceedings{roscon,
   title = {On Use of SLAM Toolbox},
   author = {{Macenski}, S.},
   year = {2019},
   journal = {ROSCon},
+  doi = {10.36288/roscon2019-900903}
 }
 @inproceedings{cartographer,
 	title = {Real-Time Loop Closure in 2D LIDAR SLAM},
 	author = {{Hess }, W. and {Kohler}, D. and {Rapp}, H. and {Andor}, D.},
 	year = {2016},
 	booktitle = {2016 IEEE International Conference on Robotics and Automation (ICRA)},
-	pages = {1271--1278}
+	pages = {1271--1278},
+	doi = {10.1109/icra.2016.7487258}
 }
 @book{thrun2005probabilistic,
 	author = {{Thrun}, S. and {Burgard}, W. and {Fox}, D.},
@@ -109,12 +112,14 @@
 	volume = {23},
 	number = {1},
 	pages = {34-46},
-	url = {http://www.ros.org/wiki/gmapping}
+	url = {http://www.ros.org/wiki/gmapping},
+	doi = {10.1109/tro.2006.889486}
 }
 @inproceedings{hector,
 	author = {{Kohlbrecher}, S. and {von Stryk}, O. and {Meyer}, J. and {Klingauf}, U.},  booktitle = {2011 IEEE International Symposium on Safety, Security, and Rescue Robotics},
 	title = {A flexible and scalable SLAM system with full 3D motion estimation},
 	year = {2011},
 	url={http://www.ros.org/wiki/hector_slam},
-	pages={155-160}
+	pages={155-160},
+	doi = {10.1109/ssrr.2011.6106777}
 }
